High scores integration

+++ core/src/info/quadtree/ld40/HighScoreEntry.java
@@ -1,7 +1,7 @@
 package info.quadtree.ld40;
 
 public class HighScoreEntry {
-    public Float Score;
+    public Integer Score;
 
     @Override
     public String toString() {
